pyproject.toml
[project]
name = "session1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "pillow>=11.3.0",
    "scipy>=1.15.3",
    "streamlit>=1.48.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
